+++ ndd/exceptions.py
@@ -4,7 +4,6 @@
 """Base EntropyEstimator class."""
 import logging
 
-
 logger = logging.getLogger(__name__)